llm/pyproject.toml
[project]
name = "llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.1",
    "aiosignal==1.4.0",
    "altair==5.5.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "asttokens==3.0.0",
    "attrs==25.4.0",
    "backoff==2.2.1",
    "bcrypt==5.0.0",
    "blinker==1.9.0",
    "blockbuster==1.5.25",
    "brotli==1.2.0",
    "build==1.3.0",
    "cachetools==6.2.1",
    "certifi==2025.10.5",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "chromadb==1.2.1",
    "click==8.3.0",
    "cloudpickle==3.1.2",
    "coloredlogs==15.0.1",
    "comm==0.2.3",
    "cryptography==44.0.3",
    "dataclasses-json==0.6.7",
    "ddgs==9.7.1",
    "debugpy==1.8.17",
    "decorator==5.2.1",
    "distro==1.9.0",
    "dotenv>=0.9.9",
    "duckduckgo-search==8.1.1",
    "durationpy==0.10",
    "executing==2.2.1",
    "fastapi==0.121.1",
    "filelock==3.20.0",
    "filetype==1.2.0",
    "flatbuffers==25.9.23",
    "forbiddenfruit==0.1.4",
    "frozenlist==1.8.0",
    "fsspec==2025.9.0",
    "gitdb==4.0.12",
    "gitpython==3.1.45",
    "google-ai-generativelanguage==0.9.0",
    "google-api-core==2.26.0",
    "google-auth==2.41.1",
    "google-genai>=1.52.0",
    "googleapis-common-protos==1.71.0",
    "greenlet==3.2.4",
    "grpcio==1.76.0",
    "grpcio-status==1.76.0",
    "grpcio-tools==1.75.1",
    "h11==0.16.0",
    "h2==4.3.0",
    "hf-xet==1.1.10",
    "hpack==4.1.0",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "huggingface-hub==0.35.3",
    "humanfriendly==10.0",
    "hyperframe==6.1.0",
    "idna==3.11",
    "importlib-metadata==8.7.0",
    "importlib-resources==6.5.2",
    "ipykernel==7.0.1",
    "ipython==9.6.0",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.25.1",
    "jsonschema-rs==0.29.1",
    "jsonschema-specifications==2025.9.1",
    "jupyter-client==8.6.3",
    "jupyter-core==5.9.1",
    "kubernetes==34.1.0",
    "langchain==1.0.2",
    "langchain-chroma==1.0.0",
    "langchain-classic==1.0.0",
    "langchain-community==0.4",
    "langchain-core==1.0.0",
    "langchain-google-genai==3.0.0",
    "langchain-ollama>=1.0.0",
    "langchain-tavily>=0.2.13",
    "langchain-text-splitters==1.0.0",
    "langgraph==1.0.1",
    "langgraph-api==0.5.12",
    "langgraph-checkpoint==3.0.1",
    "langgraph-cli[inmem]==0.4.7",
    "langgraph-prebuilt==1.0.1",
    "langgraph-runtime-inmem==0.18.0",
    "langgraph-sdk==0.2.9",
    "langsmith==0.4.37",
    "lxml==6.0.2",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "marshmallow==3.26.1",
    "matplotlib-inline==0.2.1",
    "mdurl==0.1.2",
    "mmh3==5.2.0",
    "mpmath==1.3.0",
    "multidict==6.7.0",
    "mypy-extensions==1.1.0",
    "narwhals==2.11.0",
    "nest-asyncio==1.6.0",
    "numpy==2.3.4",
    "oauthlib==3.3.1",
    "onnxruntime==1.23.2",
    "opentelemetry-api==1.38.0",
    "opentelemetry-exporter-otlp-proto-common==1.38.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.38.0",
    "opentelemetry-exporter-otlp-proto-http==1.38.0",
    "opentelemetry-proto==1.38.0",
    "opentelemetry-sdk==1.38.0",
    "opentelemetry-semantic-conventions==0.59b0",
    "orjson==3.11.3",
    "ormsgpack==1.12.0",
    "overrides==7.7.0",
    "packaging==25.0",
    "pandas==2.3.3",
    "parso==0.8.5",
    "pexpect==4.9.0",
    "pillow==12.0.0",
    "platformdirs==4.5.0",
    "posthog==5.4.0",
    "primp==0.15.0",
    "prompt-toolkit==3.0.52",
    "propcache==0.4.1",
    "proto-plus==1.26.1",
    "protobuf==6.33.0",
    "psutil==7.1.1",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pyarrow==21.0.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pybase64==1.4.2",
    "pycparser==2.23",
    "pydantic==2.12.3",
    "pydantic-core==2.41.4",
    "pydantic-settings==2.11.0",
    "pydeck==0.9.1",
    "pygments==2.19.2",
    "pyjwt==2.10.1",
    "pypika==0.48.9",
    "pyproject-hooks==1.2.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "pyzmq==27.1.0",
    "referencing==0.37.0",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rich==14.2.0",
    "rpds-py==0.27.1",
    "rsa==4.9.1",
    "setuptools==80.9.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "socksio==1.0.0",
    "sqlalchemy==2.0.44",
    "sse-starlette==2.1.3",
    "stack-data==0.6.3",
    "starlette==0.49.3",
    "streamlit==1.51.0",
    "structlog==25.5.0",
    "sympy==1.14.0",
    "tenacity==9.1.2",
    "tokenizers==0.22.1",
    "toml==0.10.2",
    "tornado==6.5.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "truststore==0.10.4",
    "typer==0.20.0",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.2",
    "urllib3==2.3.0",
    "uvicorn==0.38.0",
    "uvloop==0.22.1",
    "watchdog==6.0.0",
    "watchfiles==1.1.1",
    "wcwidth==0.2.14",
    "websocket-client==1.9.0",
    "websockets==15.0.1",
    "xxhash==3.6.0",
    "yarl==1.22.0",
    "zipp==3.23.0",
    "zstandard==0.25.0",
]
